split tests into different modules for easier maintenance
