+++ testapp/tests/test_generic.py
@@ -650,132 +650,3 @@ class Test(TestCase):
         self.assertEqual(reverse('dkron:proxy') + 'whatever', r.headers['location'])
         # some headers from upstream are dropped
         self.assertIsNone(r.headers.get('Content-Encoding'))
-
-
-@override_settings(DKRON_PATH='/dkron/proxy/ui/', DKRON_URL='http://dkron', DKRON_NAMESPACE='wtv')
-class TestNamespace(Test):
-    def test_sync_job(self, job_prefix=''):
-        super().test_sync_job(job_prefix='wtv_')
-
-    def test_delete_job(self, job_prefix=''):
-        super().test_delete_job(job_prefix='wtv_')
-
-    @mock.patch('requests.get')
-    def test_resync_jobs(self, mp1):
-        j1 = models.Job.objects.create(name='job1')
-        j2 = models.Job.objects.create(name='job2')
-        jobs = [
-            {'name': 'wtv_job1', 'tags': {'label': 'testapp'}, 'metadata': {'cron': 'auto'}},
-            {'name': 'wtv_job3', 'tags': {'label': 'testapp'}, 'metadata': {'cron': 'auto'}},
-            # ignore, not wtv namespace
-            {'name': 'ign_job4', 'tags': {'label': 'testapp'}, 'metadata': {'cron': 'auto'}},
-            # skipped, wrong label
-            {'name': 'wtv_job5', 'tags': {'label': 'testapp2'}, 'metadata': {'cron': 'auto'}},
-        ]
-        mp1.return_value = mock.MagicMock(status_code=200, json=lambda: jobs)
-        it = utils.resync_jobs()
-
-        with mock.patch('dkron.utils.sync_job') as mp2, mock.patch('dkron.utils.delete_job') as mp3:
-            el = next(it)
-            mp1.assert_called_once_with('http://dkron/v1/jobs', params={'metadata[cron]': 'auto'})
-            self.assertEqual(('job1', 'u', None), el)
-            mp2.assert_called_once_with(j1, jobs[0])
-            mp2.reset_mock()
-
-            mp2.side_effect = utils.DkronException(666, 'looking for d/a/emon')
-            el = next(it)
-            self.assertEqual(('job2', 'u', 'looking for d/a/emon'), el)
-            mp2.assert_called_once_with(j2, False)
-            mp2.reset_mock()
-
-            # no order in a set(), check for both
-            el = next(it)
-            self.assertEqual(('job3', 'd', None), el)
-            mp3.assert_called_once_with('job3')
-            mp3.reset_mock()
-
-            with self.assertRaises(StopIteration):
-                # assert nothing left
-                next(it)
-
-    def test_webhook(self, job_prefix=''):
-        super().test_webhook(job_prefix='wtv_')
-
-    def test_webhook_namespace_mismatch(self):
-        j = models.Job.objects.create(name='job1')
-        self.assertFalse(j.last_run_success)
-
-        r = self.client.post(reverse('dkron_api:webhook'), data='test\njob1\ntrue', content_type='not_form_data')
-        self.assertEqual(r.status_code, 404)
-
-        r = self.client.post(reverse('dkron_api:webhook'), data='test\nwtv_job1\ntrue', content_type='not_form_data')
-        self.assertEqual(r.status_code, 200)
-        j.refresh_from_db()
-        self.assertTrue(j.last_run_success)
-
-    def test_webhook_notify(self, job_prefix=''):
-        super().test_webhook_notify(job_prefix='wtv_')
-
-    def test_run_async(self, job_prefix=''):
-        super().test_run_async(job_prefix='wtv_')
-
-
-class TestOthers(TestCase):
-    @mock.patch('platform.machine')
-    @mock.patch('platform.system')
-    @mock.patch('requests.get')
-    @override_settings(DKRON_BIN_DIR=None)
-    def test_run_dkron_download(self, req_mock, sys_mock, mach_mock):
-        from dkron.management.commands import run_dkron
-        from io import BytesIO
-        import base64
-
-        fake_tar = base64.decodebytes(
-            # tarfile with an empty file
-            b'H4sICJNVO2IAA2EudGFyAEtkoD0wMDAwMzFRANHmZqZg2sAIwocBBUMTI0MzM1MjM0NDBQNDQzNjIwYFAzq4jaG0uCSxCOiUtPyCzLxE3OqAytLS8JgD9QecHgWjYBSMgkEOAKshHuMABgAA'
-        )
-        cmd = run_dkron.Command()
-
-        def _check(system, machine, url):
-            req_mock.reset_mock()
-            sys_mock.return_value = system
-            mach_mock.return_value = machine
-            req_mock.return_value.__enter__.return_value = mock.MagicMock(raw=BytesIO(fake_tar))
-            cmd.download_dkron()
-            req_mock.assert_called_once_with(
-                url,
-                stream=True,
-            )
-
-        _check(
-            'darwin',
-            'x86_64',
-            'https://github.com/distribworks/dkron/releases/download/v3.1.10/dkron_3.1.10_darwin_amd64.tar.gz',
-        )
-        _check(
-            'windows',
-            'AMD64',
-            'https://github.com/distribworks/dkron/releases/download/v3.1.10/dkron_3.1.10_windows_amd64.tar.gz',
-        )
-        _check(
-            'Linux',
-            'aarch64',
-            'https://github.com/distribworks/dkron/releases/download/v3.1.10/dkron_3.1.10_linux_arm64.tar.gz',
-        )
-
-    def test_utils_dkron_version_tuple(self):
-        utils.dkron_binary_version.cache_clear()
-        with override_settings(DKRON_VERSION='3.1.1'):
-            self.assertEqual(utils.dkron_binary_version(), (3, 1, 1))
-
-        utils.dkron_binary_version.cache_clear()
-        with override_settings(DKRON_VERSION='4.1.1-fork1'):
-            self.assertEqual(utils.dkron_binary_version(), (4, 1, 1))
-
-        utils.dkron_binary_version.cache_clear()
-        with override_settings(DKRON_VERSION='prefixed-4.1.2-fork1'):
-            self.assertEqual(utils.dkron_binary_version(), (4, 1, 2))
-
-        utils.dkron_binary_version.cache_clear()
-        with override_settings(DKRON_VERSION='4.1-bad'):
-            self.assertEqual(utils.dkron_binary_version(), utils.UNKNOWN_DKRON_VERSION)
